requirements.txt
# Official quick-start dependencies (lightweight path)
numpy>=1.24.0
Pillow>=10.0.0
